pyproject.toml
[project]
name = "ctfbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "openai>=1.63.2",
    "python-dotenv>=1.0.1",
]
